QDoc: Remove superfluous code comment
At some point, someone probably had an idea they wanted to pursue.
As that time is now several years ago, I'd be willing to bet what
that idea was is long lost knowledge. Drop the comments as they keep
triggering the "what was the intention here?" question, which
cannot be answered satisfactorily.

Change-Id: I2996e98ee125ed0b198ee90677de82b5986cf5a6
Reviewed-by: Luca Di Sera <luca.disera@qt.io>