QDoc: Move `CodeParser::checkModuleInclusion` to "cppcodeparser.cpp"
As part of extracting user-provided documentation, QDoc has to parse a
variety of source files in a variety of languages.

`CodeParser` is the base class of objects that take care of performing
this parsing process in QDoc.

When extracting the documentation in one of the `CodeParser`s QDoc
generally tries to parse the extracted documentation in-place.

For certain documentable elements, QDoc expects that an "\inmodule"
commands is provided, as it requires it to understand where to position
them when generating the output documentation.

To ensure that this constraint is satisfied, CppCodeParser`, the
`CodeParser` subclass that takes care of providing the semantic for QDoc
block-comments, calls at a certain point
`CodeParser::checkModuleInclusion`, a method that makes sure that the
internal representation of the comment-block and the documented element
have a valid "\inmodule" information for the later stages.

`checkModuleInclusion` is only called by `CppCodeParser` and does not
require any instance-state that would not otherwise be available.

Hence, `checkModuleInclusion` is removed from `CodeParser` in favor of
being a static in "cppcodeparser.cpp", just above its actual usage.

An access to `m_qdb`, an instance-member for `CodeParser` that reference
the `QDocDatabase` singleton was modified to directly obtain a
`QDocDatabase` instance, as the member is not available outside
`CodeParser`s.

Change-Id: I0a8abc53d3c4fbd3e6687254a0034a3e12a6e018
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>