QDoc: Extract singleExecutionMode function from main()
In QDoc's main(), two identical conditional checks for whether QDoc runs
in single execution mode follow each other. This patch extracts the
contents of these conditionals and places them in a separate function,
singleExecutionMode(). This in turn makes main() slightly easier to read
and reason about.

The new function is documented such that its purpose can be understood
in clear terms. The list the function operates on is obtained from Config.
As that is a singleton object, the function doesn't require the passing
of a reference to that object, as it can access the required functionality
directly through the singleton instance itself. The new function is
therefore parameterless.

Task-number: QTBUG-71176
Change-Id: I49fb571b99a24ff70ab8f0b17de8e1f44064e418
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>