QDoc: Extract dualExecutionMode function from main()
Recent refacoring extracted the executing code in QDoc's main() function
for running QDoc in "single execution mode". This patch does the same
for "dual execution mode", which is QDoc's default behavior. The result
is a more expressive main() function, that is easier to reason about for
code readers.

Task-number: QTBUG-71176
Change-Id: I494f58cd0d424f4cde47cc0363f908f85aa44eac
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>