QDoc: Extract helper function in main.cpp
Recent changes have extracted helper functions for running QDoc in
single or dual execution mode, respectively. Both functions do minor
setup, including getting a list of .qdocconf-files from `Config`. The
setup and list of .qdocconf-files differ in these two execution modes.

Both functions then loops over their list of .qdocconf-files. This loop
calls clear() on Config's list of module dependencies before dispatching
the qdocconf file to processQdocconfFile() for further processing. This
loop is identical to both modes of execution; in single execution mode,
this particular loop is performed twice.

This patch extracts this loop as a separate function that accepts a list
of .qdocconf-files it will iterate over. As this loop makes out the bulk
of code in the two functions for execution modes, this implementation
detail obscures the material difference of the work performed in the two
callers of this new function. By extracting it, code readability is
improved by letting the interesting parts take center stage.

Task-number: QTBUG-71176
Change-Id: Ie7f63bde17d9b676f9755874537af6c5bb1f9c7e
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>