QDoc: tiny cleanup in main.cpp
Since recent refactoring removed the main use of the qdocFiles
local variable, drop it in favor if accessing the source for
that information directly through Config::instance(). Do the same
for two other accesses through the local reference to Config::instance,
thus allowing the removal of that reference.

Task-number: QTBUG-71176
Change-Id: Id64d58cacc3b368d1986e12119a71cdbdc8d648a
Reviewed-by: Luca Di Sera <luca.disera@qt.io>