qdoc: Warn about documented global functions that generate no docs
Generator::generateDocumentation() recursively traverses the tree of
PageNode objects, i.e. nodes that generate output. However, it did
not check if there are documented nodes in the global scope (root
namespace) that generate no output.

A typical case where this happens is global functions, where the
author forgot to associate the function to something using the
\relates command.

Make QDoc print out a warning for these nodes, and mark them
to have 'DontDocument' status so they will be ignored later on
for linking and inclusion in the .qhp file.

Also in Generator::generateDocumentation(), rename a variable local
to a for-loop from 'node' to 'child', to prevent shadowing a 'node'
in the outer scope.

Fixes: QTBUG-112256
Change-Id: I74fcc704f6848fc1eef8529da80f57678a83766e
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 65e94d012b7d38669e9cdc4e885bdc34a86607af)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>