QDoc: Remove unused variable `sourceFileNames`
When QDoc runs it collects a series of input source files based on
certain configuration variables.

The files are collected into two maps, in `processQDocConfFile`,
`sources` and `sourceFileNames`.

The collected input source files are later used as the base for
extracting the user-provided documentation.

Nonetheless, `sourceFileNames`, which is populated along with `sources`
was never used and is thus removed as dead code.

Change-Id: Ie02025e8b1054e4a4f3e9235ee1cc2d7631a2935
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>