QDoc: Remove nesting level from `ClangCodeParser::buildPCH`
`ClangCodeParser`, the class responsible for parsing Cpp source files
during a QDoc run, uses a PCH to speed up compilation when reading the
above mentioned source file.

A PCH file is built in `ClangCodeParser::buildPCH`.
To avoid redoing the same work more than one time if it is unnecessary,
`buildPCH` branches on certain conditions and does nothing if those
conditions are not met.

To make the code easier to read, move the code that performs the PCH
building operation to the top level of the method body, instead of
branching, and use early returns when the conditions are not satisfied.

Change-Id: I95b07a12bc1723ea7bdaee41367483a61dcb5609
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>