Allow to opt out of source path checks
Commit 454ab5ffd5b07 ff let qtattributionsscanner bail out if the
content of the "Path" and "Files" properties are not valid relative
file paths at the time qtattributionsscanner runs.

For Qt for MCU tough, the relative path to third-party files does
differ between the git checkout, and the source SDK as given to the
user.

This patch, therefore, allows opting out of the check by adding a
--no-path-checks option.

Pick-to: 6.5
Change-Id: I809b52e233a6b56194feb1b5b8168f36ba1b9972
Reviewed-by: Venugopal Shivashankar <Venugopal.Shivashankar@qt.io>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>