lupdate: Fix assert when applying number heuristics
There's an edge case in the number heuristics of lupdate that triggers a
Q_ASSERT due to an out-of-bounds string character access. To run into
the assert, the following must be true:

- oldSource == oldTranslation
- oldSource must contain a number followed immediately by more than one
  characters
- oldSource and newSource must differ only in the character after the
  number

Examples that trigger the assert:

| oldSource  | newSource  |
|------------+------------|
| "%1MiB"    | "%1Kib"    |
| "1Mi"      | "1Ki"      |
| "1M "      | "1Ki "     |
| "123abcde" | "123xbcde" |

Examples that do not trigger the assert:

| oldSource  | newSource  |
|------------+------------|
| "%1MiB"    | "%1Kob"    |
| "1M"       | "1K"       |
| "123abcde" | "123xxcde" |

Add a test case that triggers the Q_ASSERT.

Add a check to avoid the out-of-bounds access as a minimal fix. The
whole number heuristics algorithm is fishy and produces results that
don't match the source code comments. But it's been like that "forever",
so let's not change existing behavior people may rely on.

Fixes: QTBUG-111775
Change-Id: Ibb6bcce9115f7060b6de3f97451e38105623c003
Reviewed-by: Kai Köhne <kai.koehne@qt.io>
(cherry picked from commit 811b7c2f7eaac22e64cbb92508b67ccad5eed07c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>