QDoc: Remove unused method from FunctionNode
The FunctionNode class contains the public method setVirtual().
The method's intent is to set the private member m_virtualness
to FunctionNode::Virtualness::NormalVirtual. This behavior is
achieved, and exercised in QDoc, by calling another method of
the same class, setVirtualness(Virtualness), which offers the
option of passing any value from the FunctionNode::Virtualness
enum, making that API more flexible than that of setVirtual()
which doesn't take any arguments.

As the private member is set in code by a different setter,
and because the setVirtual() method is unused, remove the latter
as dead code.

Task-number: QTBUG-71176
Change-Id: Ieab1fa5b6be92e481fc9df887816fec3f08bbdaf
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>