QDoc: Ensure FunctionNode sets sensible virtualness values
The index files generated by QDoc contains an attribute for a
function's "virtualness"; that is, if the function is a virtual, pure
virtual, or not virtual. These attributes are written to the index
file if they exist, and read back in, again if they exist. For
functions that do not have this attribute in the index file, a
virtualness of NonVirtual is implicitly assumed.

This change explicitly initializes the responsible member variable,
FunctionNode::m_virtualness, to NonVirtual. Furthermore, the setter
overload setVirtualness(const QString&), which is called only when
reading an index file back in, is modified to handle the special case
for pure virtuals first, then setting m_virtualness to Virtual if
that's the case, and NonVirtual for any empty, missing, or unknown
attributes.

Task-number: QTBUG-112494
Change-Id: I8f08b72ae58dca5da07932f010da230c667b056c
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>