Doc: Use the \examplecategory macro to tag examples
The '\meta category' command was used for tagging examples with a
specific category, used in Qt Creators Welcome mode.

As we want to also generate lists of examples belonging to a category
elsewhere in the documentation, replace the command with a macro that
expands to the original \meta command and also adds the example to
a group using the \ingroup command. This way, the category names can
be used as arguments to the \generatelist or \annotatedlist commands.

Task-number: QTBUG-112731
Change-Id: I28dfd6d5540dfe540bab942f9e9a93e9c239e206
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit a3379894bba777d4dcc8dd6ced5149533bb127ed)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>