QDoc: Remove unreachable code in WebXMLGenerator::addAtomElements
When the WebXML generator was revived in 2017, a case for handling QML
atoms in the method `WebXMLGenerator::addAtomElements` was added under
an `#ifdef QDOC_QML` guard. This is, according to esteemed historians,
a remnant from the old Qt 4.7 implementation of the WebXML generator.
However, since the revival, the define wasn't ever set by anything in
neither QDoc nor the build system, effectively rendering this code
unreachable.

This change removes the `#ifdef [...] #endif` block, as removing the
unused code won't lead to behavior change of QDoc.

Task-number: QTBUG-112764
Change-Id: Ib9b50d4831e86c60de05065d315f421ae2289f64
Reviewed-by: Luca Di Sera <luca.disera@qt.io>