QDoc: Refactor DocParser::insertTarget
A recent refactoring extracted a method from `DocParser::insertTarget`
that allows and early return. This makes the else branch redundant, as
control will always exit the method if the `if` condition is satisfied.
This allows for flattening the method further.

This change thus flattens the method by removing the else statement and
flattening the code in that branch. The remaining code features another
if/else check. This change adds a return statement to the if statement,
allowing the removal of the following else statement and the flattening
of that branch, too.

Task-number: QTBUG-113126
Change-Id: If0e7436f8bb98924dd54c9d9f04a8a7a4d7566a9
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>