Revert "Check the return value from QTranslator::load()"
This reverts commit 0a91ac627e84634993fc579d428e4bbc3d0b283a.

The example documentation does not match the flow introduced
by this change. The example explicitly talks about "No error
will occur if the file is not found.", and "If hellotr_la.qm
was found and contains a translation for "Hello world!", the
translation appears; if not, the source text appears."

Q_UNUSED added to satisfy nodiscard attribute.

Pick-to: 6.5
Change-Id: I8e4cb6ed3fb81f7739aacd734f3107a8ec6899a1
Reviewed-by: Kai Köhne <kai.koehne@qt.io>