QDoc: Extract DocParser::insertKeyword from insertTarget
Recent refactoring of `DocParser::insertTarget` exposed that the method
is responsible for the handling of two separate concepts in QDoc, the
commands `\target` and `\keyword`, respectively. Handling of the latter
command was added to the method in question by extending it's signature
with a boolean parameter, whose value was significant in deciding which
code path to follow. However, this API design is no longer considered
appropriate for QDoc, as discussed in review of the previous
refactoring.

This patch extracts the code responsible for inserting keywords into
their related data structure from `DocParser::insertTarget`, and adds
this code to a new member method, `DocParser::insertKeyword`. The
signature of the former method is updated and the relevant call sites
are updated to call the correct method.

The new method is documented as part of its introduction to the code
base. Given that the two methods have striking similarities,
documentation is added to `DocParser::insertTarget` in its new form.

Fixes: QTBUG-113126
Change-Id: Id29bc655c15b6e12726100eb1f33adbf97487236
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>